requirements.txt
# Custom Django model-views-forms requirements
Django==1.8.12
django-extra-views
django-cuser==2014.9.28
django-widget-tweaks==1.4.1
python-magic

# Documentation requirements
sphinx
sphinx_rtd_theme

# Custom Report helpers requirements
openpyxl==2.2.3
xlwt

# Fabric recipes requirements
Fabric

# Testing helpers requirements
pyvirtualdisplay
requests
splinter
django-behave

#CCE Toolkit
git+https://github.com/cceit/cce-toolkit.git@latest

#Postgres
psycopg2==2.6.1

Pillow
django-countries==5.0
